Move device util from gaen module
