+++ src/Device/index.ts
@@ -1,0 +1,3 @@
+import { openAppSettings } from "./nativeModule"
+
+export { openAppSettings }
