+++ src/Device/nativeModule.ts
@@ -1,0 +1,8 @@
+import { NativeModules } from "react-native"
+
+// Utils Module
+const utilsModule = NativeModules.UtilsModule
+
+export const openAppSettings = (): void => {
+  utilsModule.openAppSettings()
+}
